+++ src/components/EditionCard/CardRequiredBadge.tsx
@@ -3,8 +3,8 @@ import { StatusBadge } from "@nypl/design-system-react-components";
 
 const CardRequiredBadge: React.FC = () => {
   return (
-    <StatusBadge level="high" width={{ base: "100%", md: "fit-content" }}>
-      Library Card Required
+    <StatusBadge type="warning" width={{ base: "100%", md: "fit-content" }}>
+      LIBRARY CARD REQUIRED
     </StatusBadge>
   );
 };