+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Update README to include production release information
 - Upgrade NYPL Design System to 3.1.1
 - Upgrade Web Reader to 4.3.4
+- Fix issue where 0 character is shown when there are no authors 
 
 ## [0.18.0]